Modifying workflow and namelist vars
The CPM utility has been updated so the changes are reflecting
the new namelist variables. Also, the workflow for creating the
tslist file required wrfinput_d05 which is not created until
after real.exe, so, I changed it to use a met_em file and moved
this until after the files are linked. This should make things
more streamlined.